Forecast Accuracy Metrics Sets

This is a wrapper for metric_set() with several common forecast / regression accuracy metrics included. These are the default time series accuracy metrics used with modeltime_accuracy().

default_forecast_accuracy_metric_set()

Details

The primary purpose is to use the default accuracy metrics to calculate the following forecast accuracy metrics using modeltime_accuracy():

MAE - Mean absolute error, mae()

MAPE - Mean absolute percentage error, mape()

MASE - Mean absolute scaled error, mase()

SMAPE - Symmetric mean absolute percentage error, smape()

RMSE - Root mean squared error, rmse()

RSQ - R-squared, rsq()

Examples

library(tibble)
library(dplyr)
library(timetk)

set.seed(1)
data <- tibble(
    time  = tk_make_timeseries("2020", by = "sec", length_out = 10),
    y     = 1:10 + rnorm(10),
    y_hat = 1:10 + rnorm(10)
)

# Default Metric Specification
default_forecast_accuracy_metric_set()

#>   metric          class direction
#> 1    mae numeric_metric  minimize
#> 2   mape numeric_metric  minimize
#> 3   mase numeric_metric  minimize
#> 4  smape numeric_metric  minimize
#> 5   rmse numeric_metric  minimize
#> 6    rsq numeric_metric  maximize

# Create a metric summarizer function from the metric set
calc_default_metrics <- default_forecast_accuracy_metric_set()

# Apply the metric summarizer to new data
calc_default_metrics(data, y, y_hat)

#> # A tibble: 6 x 3
#>   .metric .estimator .estimate
#>   <chr>   <chr>          <dbl>
#> 1 mae     standard       0.979
#> 2 mape    standard      71.1  
#> 3 mase    standard       0.865
#> 4 smape   standard      31.8  
#> 5 rmse    standard       1.47 
#> 6 rsq     standard       0.795