Fit a Model Specification to a Dataset

fit() and fit_xy() take a model specification, translate the required code by substituting arguments, and execute the model fit routine.

# S3 method for arima_reg
fit(object, formula, data, control = control_parsnip(), ...)

# S3 method for arima_boost
fit(object, formula, data, control = control_parsnip(), ...)

# S3 method for exp_smoothing
fit(object, formula, data, control = control_parsnip(), ...)

# S3 method for prophet_reg
fit(object, formula, data, control = control_parsnip(), ...)

# S3 method for prophet_boost
fit(object, formula, data, control = control_parsnip(), ...)

# S3 method for seasonal_decomp
fit(object, formula, data, control = control_parsnip(), ...)

Arguments

object
An object of class model_spec that has a chosen engine (via set_engine()).

formula
An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.

data
Optional, depending on the interface (see Details below). A data frame containing all relevant variables (e.g. outcome(s), predictors, case weights, etc). Note: when needed, a named argument should be used.

control
A named list with elements verbosity and catch. See control_parsnip().

...
Not currently used; values passed here will be ignored. Other options required to fit the model should be passed using set_engine().

Value

A model_fit object that contains several elements:

lvl: If the outcome is a factor, this contains the factor levels at the time of model fitting.

spec: The model specification object (object in the call to fit)

fit: when the model is executed without error, this is the model object. Otherwise, it is a try-error object with the error message.

preproc: any objects needed to convert between a formula and non-formula interface (such as the terms object)

The return value will also have a class related to the fitted model (e.g. "_glm") before the base class of "model_fit".

Details

fit() and fit_xy() substitute the current arguments in the model specification into the computational engine's code, checks them for validity, then fits the model using the data and the engine-specific code. Different model functions have different interfaces (e.g. formula or x/y) and these functions translate between the interface used when fit() or fit_xy() were invoked and the one required by the underlying model.

When possible, these functions attempt to avoid making copies of the data. For example, if the underlying model uses a formula and fit() is invoked, the original data are references when the model is fit. However, if the underlying model uses something else, such as x/y, the formula is evaluated and the data are converted to the required format. In this case, any calls in the resulting model objects reference the temporary objects used to fit the model.

If the model engine has not been set, the model's default engine will be used (as discussed on each model page). If the verbosity option of control_parsnip() is greater than zero, a warning will be produced.

Examples

# TODO