Calculate Accuracy Metrics

This is a wrapper for yardstick that simplifies time series regression accuracy metric calculations from a fitted workflow (trained workflow) or model_fit (trained parsnip model).

modeltime_accuracy(
  object,
  new_data = NULL,
  metric_set = default_forecast_accuracy_metric_set(),
  acc_by_id = FALSE,
  quiet = TRUE,
  ...
)

Arguments

object
A Modeltime Table

new_data
A tibble to predict and calculate residuals on. If provided, overrides any calibration data.

metric_set
A yardstick::metric_set() that is used to summarize one or more forecast accuracy (regression) metrics.

acc_by_id
Should a global or local model accuracy be produced? (Default: FALSE)

When FALSE, a global model accuracy is provided.

If TRUE, a local accuracy is provided group-wise for each time series ID. To enable local accuracy, an id must be provided during modeltime_calibrate().

quiet
Hide errors (TRUE, the default), or display them as they occur?

...
If new_data is provided, these parameters are passed to modeltime_calibrate()

Value

A tibble with accuracy estimates.

Details

The following accuracy metrics are included by default via default_forecast_accuracy_metric_set():

MAE - Mean absolute error, mae()

MAPE - Mean absolute percentage error, mape()

MASE - Mean absolute scaled error, mase()

SMAPE - Symmetric mean absolute percentage error, smape()

RMSE - Root mean squared error, rmse()

RSQ - R-squared, rsq()

Examples

library(tidymodels)

#> -- Attaching packages -------------------------------------- tidymodels 0.1.3 --

#> v broom        0.7.9      v tibble       3.1.3 
#> v dials        0.0.10     v tidyr        1.1.3 
#> v ggplot2      3.3.5      v tune         0.1.6 
#> v infer        1.0.0      v workflows    0.2.3 
#> v modeldata    0.1.1      v workflowsets 0.1.0 
#> v purrr        0.3.4      v yardstick    0.0.8 
#> v recipes      0.1.17     

#> Warning: package ‘broom’ was built under R version 4.0.5

#> Warning: package ‘dials’ was built under R version 4.0.5

#> Warning: package ‘ggplot2’ was built under R version 4.0.5

#> Warning: package ‘infer’ was built under R version 4.0.5

#> Warning: package ‘modeldata’ was built under R version 4.0.5

#> Warning: package ‘recipes’ was built under R version 4.0.5

#> Warning: package ‘tibble’ was built under R version 4.0.5

#> Warning: package ‘tidyr’ was built under R version 4.0.5

#> Warning: package ‘tune’ was built under R version 4.0.5

#> Warning: package ‘workflows’ was built under R version 4.0.5

#> Warning: package ‘workflowsets’ was built under R version 4.0.5

#> Warning: package ‘yardstick’ was built under R version 4.0.5

#> -- Conflicts ----------------------------------------- tidymodels_conflicts() --
#> x purrr::discard() masks scales::discard()
#> x dplyr::filter()  masks stats::filter()
#> x purrr::is_null() masks testthat::is_null()
#> x dplyr::lag()     masks stats::lag()
#> x tidyr::matches() masks rsample::matches(), dplyr::matches(), testthat::matches()
#> x tidyr::spread()  masks greybox::spread()
#> x recipes::step()  masks stats::step()
#> * Use tidymodels_prefer() to resolve common conflicts.

library(tidyverse)

#> -- Attaching packages --------------------------------------- tidyverse 1.3.1 --

#> v readr   2.0.2     v forcats 0.5.1
#> v stringr 1.4.0     

#> Warning: package ‘readr’ was built under R version 4.0.5

#> Warning: package ‘forcats’ was built under R version 4.0.3

#> -- Conflicts ------------------------------------------ tidyverse_conflicts() --
#> x readr::col_factor()    masks scales::col_factor()
#> x purrr::discard()       masks scales::discard()
#> x readr::edition_get()   masks testthat::edition_get()
#> x dplyr::filter()        masks stats::filter()
#> x stringr::fixed()       masks recipes::fixed()
#> x purrr::is_null()       masks testthat::is_null()
#> x dplyr::lag()           masks stats::lag()
#> x readr::local_edition() masks testthat::local_edition()
#> x tidyr::matches()       masks rsample::matches(), dplyr::matches(), testthat::matches()
#> x readr::spec()          masks yardstick::spec()
#> x tidyr::spread()        masks greybox::spread()

library(lubridate)

#> Warning: package ‘lubridate’ was built under R version 4.0.5

#> 
#> Attaching package: ‘lubridate’

#> The following object is masked from ‘package:greybox’:
#> 
#>     hm

#> The following objects are masked from ‘package:base’:
#> 
#>     date, intersect, setdiff, union

library(timetk)

# Data
m750 <- m4_monthly %>% filter(id == "M750")

# Split Data 80/20
splits <- initial_time_split(m750, prop = 0.8)

# --- MODELS ---

# Model 1: prophet ----
model_fit_prophet <- prophet_reg() %>%
    set_engine(engine = "prophet") %>%
    fit(value ~ date, data = training(splits))

#> Error in prophet_reg(): could not find function "prophet_reg"

# ---- MODELTIME TABLE ----

models_tbl <- modeltime_table(
    model_fit_prophet
)

#> Error in modeltime_table(model_fit_prophet): could not find function "modeltime_table"

# ---- ACCURACY ----

models_tbl %>%
    modeltime_calibrate(new_data = testing(splits)) %>%
    modeltime_accuracy(
        metric_set = metric_set(mae, rmse, rsq)
    )

#> Error in modeltime_accuracy(., metric_set = metric_set(mae, rmse, rsq)): could not find function "modeltime_accuracy"