Recursive Model Predictions

Make predictions from a recursive model.

# S3 method for recursive
predict(object, new_data, type = NULL, opts = list(), ...)

Arguments

object
An object of class model_fit

new_data
A rectangular data object, such as a data frame.

type
A single character value or NULL. Possible values are "numeric", "class", "prob", "conf_int", "pred_int", "quantile", "time", "hazard", "survival", or "raw". When NULL, predict() will choose an appropriate value based on the model's mode.

opts
A list of optional arguments to the underlying predict function that will be used when type = "raw". The list should not include options for the model object or the new data being predicted.

...
Arguments to the underlying model's prediction function cannot be passed here (see opts). There are some parsnip related options that can be passed, depending on the value of type. Possible arguments are:

level: for types of "conf_int" and "pred_int" this is the parameter for the tail area of the intervals (e.g. confidence level for confidence intervals). Default value is 0.95.

std_error: add the standard error of fit or prediction (on the scale of the linear predictors) for types of "conf_int" and "pred_int". Default value is FALSE.

quantile: the quantile(s) for quantile regression (not implemented yet)

time: the time(s) for hazard and survival probability estimates.

Value

Numeric values for the recursive panel prediction

Details

Refer to recursive() for further details and examples.