Update the model by model id in a Modeltime Table

Update the model by model id in a Modeltime Table

Usage

update_modeltime_model(object, .model_id, .new_model)

Arguments

object

A Modeltime Table

.model_id

A numeric value matching the .model_id that you want to update

.new_model

A fitted workflow, model_fit, or mdl_time_ensmble object

See also

Examples

# \donttest{
library(tidymodels)

model_fit_ets <- exp_smoothing() %>%
    set_engine("ets") %>%
    fit(value ~ date, training(m750_splits))
#> frequency = 12 observations per 1 year

m750_models %>%
    update_modeltime_model(1, model_fit_ets)
#> # Modeltime Table
#> # A tibble: 3 × 3
#>   .model_id .model     .model_desc
#>       <int> <list>     <chr>      
#> 1         1 <fit[+]>   ETS(A,A,A) 
#> 2         2 <workflow> PROPHET    
#> 3         3 <workflow> GLMNET     
# }