Tuning Parameters for ADAM Models

Tuning Parameters for ADAM Models

Usage

use_constant(values = c(FALSE, TRUE))

regressors_treatment(values = c("use", "select", "adapt"))

outliers_treatment(values = c("ignore", "use", "select"))

probability_model(
  values = c("none", "auto", "fixed", "general", "odds-ratio", "inverse-odds-ratio",
    "direct")
)

distribution(
  values = c("default", "dnorm", "dlaplace", "ds", "dgnorm", "dlnorm", "dinvgauss",
    "dgamma")
)

information_criteria(values = c("AICc", "AIC", "BICc", "BIC"))

select_order(values = c(FALSE, TRUE))

Arguments

values

A character string of possible values.

Value

A dials parameter

A parameter

A parameter

A parameter

A parameter

A parameter

A parameter

A parameter

Details

The main parameters for ADAM models are:

non_seasonal_ar: The order of the non-seasonal auto-regressive (AR) terms.

non_seasonal_differences: The order of integration for non-seasonal differencing.

non_seasonal_ma: The order of the non-seasonal moving average (MA) terms.

seasonal_ar: The order of the seasonal auto-regressive (SAR) terms.

seasonal_differences: The order of integration for seasonal differencing.

seasonal_ma: The order of the seasonal moving average (SMA) terms.

use_constant: Logical, determining, whether the constant is needed in the model or not.

regressors_treatment: The variable defines what to do with the provided explanatory variables.

outliers_treatment: Defines what to do with outliers.

probability_model: The type of model used in probability estimation.

distribution: What density function to assume for the error term.

information_criteria: The information criterion to use in the model selection / combination procedure.

select_order: If TRUE, then the function will select the most appropriate order.

Examples

use_constant()
#> Warning: The `default` argument of `new_qual_param()` is deprecated as of dials 1.1.0.
#> ℹ The deprecated feature was likely used in the modeltime package.
#>   Please report the issue at
#>   <https://github.com/business-science/modeltime/issues>.
#> Logical, determining, whether the constant is needed in the model or not  (qualitative)
#> 2 possible values include:
#> FALSE and TRUE 

regressors_treatment()
#> The variable defines what to do with the provided explanatory variables.  (qualitative)
#> 3 possible values include:
#> 'use', 'select' and 'adapt' 

distribution()
#> What density function to assume for the error term.  (qualitative)
#> 8 possible values include:
#> 'default', 'dnorm', 'dlaplace', 'ds', 'dgnorm', 'dlnorm', 'dinvgauss' and 'dg...