Calculate Accuracy Metrics

This is a wrapper for yardstick that simplifies time series regression accuracy metric calculations from a fitted workflow (trained workflow) or model_fit (trained parsnip model).

modeltime_accuracy(
  object,
  new_data = NULL,
  metric_set = default_forecast_accuracy_metric_set(),
  quiet = TRUE,
  ...
)

Arguments

object
A Modeltime Table

new_data
A tibble to predict and calculate residuals on. If provided, overrides any calibration data.

metric_set
A yardstick::metric_set() that is used to summarize one or more forecast accuracy (regression) metrics.

quiet
Hide errors (TRUE, the default), or display them as they occur?

...
Not currently used

Value

A tibble with accuracy estimates.

Details

The following accuracy metrics are included by default via default_forecast_accuracy_metric_set():

MAE - Mean absolute error, mae()

MAPE - Mean absolute percentage error, mape()

MASE - Mean absolute scaled error, mase()

SMAPE - Symmetric mean absolute percentage error, smape()

RMSE - Root mean squared error, rmse()

RSQ - R-squared, rsq()

Examples

library(tidyverse)
library(lubridate)
library(timetk)
library(parsnip)
library(rsample)

# Data
m750 <- m4_monthly %>% filter(id == "M750")

# Split Data 80/20
splits <- initial_time_split(m750, prop = 0.9)

# --- MODELS ---

# Model 1: auto_arima ----
model_fit_arima <- arima_reg() %>%
    set_engine(engine = "auto_arima") %>%
    fit(value ~ date, data = training(splits))

#> frequency = 12 observations per 1 year

# ---- MODELTIME TABLE ----

models_tbl <- modeltime_table(
    model_fit_arima
)

# ---- ACCURACY ----

models_tbl %>%
    modeltime_calibrate(new_data = testing(splits)) %>%
    modeltime_accuracy(
        metric_set = metric_set(mae, rmse, rsq)
    )

#> # A tibble: 1 x 6
#>   .model_id .model_desc             .type   mae  rmse   rsq
#>       <int> <chr>                   <chr> <dbl> <dbl> <dbl>
#> 1         1 ARIMA(0,1,1)(0,1,1)[12] Test   151.  198. 0.930