Forecast future data

The goal of modeltime_forecast() is to simplify the process of forecasting future data.

modeltime_forecast(
  object,
  new_data = NULL,
  h = NULL,
  actual_data = NULL,
  conf_interval = 0.95,
  ...
)

Arguments

object
A Modeltime Table that has been calibrated with modeltime_calibrate()

new_data
A tibble containing future information to forecast. If NULL, forecasts the calibration data.

h
The forecast horizon (can be used instead of new_data for time series with no exogenous regressors). Extends the calibration data h periods into the future.

actual_data
Reference data that is combined with the output tibble and given a .key = "actual"

conf_interval
An estimated confidence interval based on the calibration data. This is designed to estimate future confidence from out-of-sample prediction error.

...
Not currently used

Value

A tibble with predictions and time-stamp data. For ease of plotting and calculations, the column names are transformed to:

.key: Values labeled either "prediction" or "actual"

.index: The timestamp index.

.value: The value being forecasted.

.conf_lo: The lower limit of the confidence interval.

.conf_hi: The upper limit of the confidence interval.

Additional descriptive columns are included:

.model_id: Model ID from the Modeltime Table

.model_desc: Model Description from the Modeltime Table

Unnecessary columns are dropped to save space:

.model

.calibration_data

Details

The key parameters are (controlled by new_data or h) and combining with existing data (controlled by actual_data) in preparation for visualization with plot_modeltime_forecast().

Specifying New Data or Horizon (h)

When forecasting you can specify future data using:

new_data: This is a future tibble with date column and columns for xregs extending the trained dates and exogonous regressors (xregs) if used.

Forecasting Evaluation Data: By default, the new_data will use the .calibration_data if new_data is not provided. This is the equivalent of using rsample::testing() for getting test data sets.

Forecasting Future Data: See future_frame() for creating future tibbles.

Xregs: Can be used with this method

h: This is a phrase like "1 year", which extends the .calibration_data into the future.

Forecasting Future Data: All forecasts using h are extended after the calibration data, which is desirable after refitting with modeltime_refit(). Internally, a call is made to future_frame() to expedite creating new data using the date feature.

Xregs: Cannot be used because future data must include new xregs.

Actual Data

This is reference data that contains the true values of the time-stamp data. It helps in visualizing the performance of the forecast vs the actual data.

Confidence Interval Estimation

Confidence intervals (.conf_lo, .conf_hi) are estimated based on the normal estimation of the testing errors (out of sample) from modeltime_calibrate(). The out-of-sample error estimates are then carried through and applied to applied to any future forecasts.

The confidence interval can be adjusted with the conf_interval parameter. An 80% confidence interval estimates a normal (Gaussian distribution) that assumes that 80% of the future data will fall within the upper and lower confidence limits.

The confidence interval is mean-adjusted, meaning that if the mean of the residuals is non-zero, the confidence interval is adjusted to widen the interval to capture the difference in means.

Refitting has no affect on the confidence interval since this is calculated independently of the refitted model (on data with a smaller sample size). New observations typically improve future accuracy, which in most cases makes the out-of-sample confidence intervals conservative.

Examples

library(tidyverse)
library(lubridate)
library(timetk)
library(parsnip)
library(rsample)

# Data
m750 <- m4_monthly %>% filter(id == "M750")

# Split Data 80/20
splits <- initial_time_split(m750, prop = 0.9)

# --- MODELS ---

# Model 1: auto_arima ----
model_fit_arima <- arima_reg() %>%
    set_engine(engine = "auto_arima") %>%
    fit(value ~ date, data = training(splits))

#> frequency = 12 observations per 1 year

# ---- MODELTIME TABLE ----

models_tbl <- modeltime_table(
    model_fit_arima
)

# ---- CALIBRATE ----

calibration_tbl <- models_tbl %>%
    modeltime_calibrate(new_data = testing(splits))

# ---- ACCURACY ----

calibration_tbl %>%
    modeltime_accuracy()

#> # A tibble: 1 x 9
#>   .model_id .model_desc             .type   mae  mape  mase smape  rmse   rsq
#>       <int> <chr>                   <chr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1         1 ARIMA(0,1,1)(0,1,1)[12] Test   151.  1.41 0.516  1.43  198. 0.930

# ---- FORECAST ----

calibration_tbl %>%
    modeltime_forecast(
        new_data    = testing(splits),
        actual_data = m750
    )

#> # A tibble: 337 x 7
#>    .model_id .model_desc .key   .index     .value .conf_lo .conf_hi
#>        <int> <chr>       <fct>  <date>      <dbl>    <dbl>    <dbl>
#>  1        NA ACTUAL      actual 1990-01-01   6370       NA       NA
#>  2        NA ACTUAL      actual 1990-02-01   6430       NA       NA
#>  3        NA ACTUAL      actual 1990-03-01   6520       NA       NA
#>  4        NA ACTUAL      actual 1990-04-01   6580       NA       NA
#>  5        NA ACTUAL      actual 1990-05-01   6620       NA       NA
#>  6        NA ACTUAL      actual 1990-06-01   6690       NA       NA
#>  7        NA ACTUAL      actual 1990-07-01   6000       NA       NA
#>  8        NA ACTUAL      actual 1990-08-01   5450       NA       NA
#>  9        NA ACTUAL      actual 1990-09-01   6480       NA       NA
#> 10        NA ACTUAL      actual 1990-10-01   6820       NA       NA
#> # ... with 327 more rows