Prepared Nested Modeltime Data

A set of functions to simplify preparation of nested data for iterative (nested) forecasting with Nested Modeltime Tables.

extend_timeseries(.data, .id_var, .date_var, .length_future)

nest_timeseries(.data, .id_var, .length_future, .length_actual = NULL)

split_nested_timeseries(.data, .length_test, .length_train = NULL, ...)

Arguments

.data
A data frame or tibble containing time series data. The data should have:

identifier (.id_var): Identifying one or more time series groups

date variable (.date_var): A date or date time column

target variable (.value): A column containing numeric values that is to be forecasted

.id_var
An id column

.date_var
A date or datetime column

.length_future
Varies based on the function:

extend_timeseries(): Defines how far into the future to extend the time series by each time series group.

nest_timeseries(): Defines which observations should be split into the .future_data.

.length_actual
Can be used to slice the .actual_data to a most recent number of observations.

.length_test
Defines the length of the test split for evaluation.

.length_train
Defines the length of the training split for evaluation.

...
Additional arguments passed to the helper function. See details.

Details

Preparation of nested time series follows a 3-Step Process:

Step 1: Extend the Time Series

extend_timeseries(): A wrapper for timetk::future_frame() that extends a time series group-wise into the future.

The group column is specified by .id_var.

The date column is specified by .date_var.

The length into the future is specified with .length_future.

Step 2: Nest the Time Series

nest_timeseries(): A helper for nesting your data into .actual_data and .future_data.

The group column is specified by .id_var

The .length_future defines the length of the .future_data.

The remaining data is converted to the .actual_data.

The .length_actual can be used to slice the .actual_data to a most recent number of observations.

The result is a "nested data frame".

Step 3: Split the Actual Data into Train/Test Splits

split_nested_timeseries(): A wrapper for timetk::time_series_split() that generates training/testing splits from the .actual_data column.

The .length_test is the primary argument that identifies the size of the testing sample. This is typically the same size as the .future_data.

The .length_train is an optional size of the training data.

The ... (dots) are additional arguments that can be passed to timetk::time_series_split().

Examples

library(tidyverse)
library(timetk)
library(modeltime)

walmart_sales_weekly %>%
    select(id, Date, Weekly_Sales) %>%
    set_names(c("id", "date", "value")) %>%

    # Extends the time series by id
    extend_timeseries(
        .id_var     = id,
        .date_var   = date,
        .length_future = 52
    ) %>%

    # Nests the time series into .actual_data and .future_data
    nest_timeseries(
        .id_var     = id,
        .length_future = 52
    ) %>%

    # Adds a column .splits that contains training/testing samples
    split_nested_timeseries(
        .length_test = 52
    )

#> # A tibble: 7 x 4
#>   id    .actual_data       .future_data      .splits        
#>   <fct> <list>             <list>            <list>         
#> 1 1_1   <tibble [143 x 2]> <tibble [52 x 2]> <split [91/52]>
#> 2 1_3   <tibble [143 x 2]> <tibble [52 x 2]> <split [91/52]>
#> 3 1_8   <tibble [143 x 2]> <tibble [52 x 2]> <split [91/52]>
#> 4 1_13  <tibble [143 x 2]> <tibble [52 x 2]> <split [91/52]>
#> 5 1_38  <tibble [143 x 2]> <tibble [52 x 2]> <split [91/52]>
#> 6 1_93  <tibble [143 x 2]> <tibble [52 x 2]> <split [91/52]>
#> 7 1_95  <tibble [143 x 2]> <tibble [52 x 2]> <split [91/52]>